Created TicTacToe.py.
The file TicTacToe contains the class TTT_Board which contains encapsulates the representation of the tictactoe board and the game logic. The interface contains only one method which is play_game(gamemode="interactive", gamefile=None) which runs the game either in interactive mode or non-interactive mode in which case it reads the gamefile for the input choices.